+++ src/helpers.js
@@ -1,5 +1,7 @@
+const _cloneDeep = require('lodash.clonedeep');
 const cheerio = require('cheerio');
 const htmlparser2 = require('htmlparser2');
+const Vue = require ('vue');
 
 const helpers = {
   /**
@@ -21,6 +23,64 @@ const helpers = {
     const dom = htmlparser2.parseDOM(html, xmlOptions);
     const $ = cheerio.load(dom, { xml: xmlOptions });
     return $;
+  },
+
+  /**
+   * Creates a Vue instance to render the vnode as an HTML string.
+   *
+   * @param  {object} vnode  Vue's vnode object
+   * @return {string}        The rendered HTML
+   */
+  vnodeToString: function (vnode) {
+    const vm = new Vue({
+      render: function () {
+        return vnode;
+      }
+    });
+    const html = vm.$mount().$el.outerHTML;
+    vm.$destroy();
+    return html;
+  },
+
+  // This does not seem to make an actual copy. It is still modifying the reference.
+  /**
+   * Makes a copy of the vnode, so we are not mutating the original reference passed in by the test.
+   *
+   * @param  {object} vnode Vue's vnode from the wrapper
+   * @return {object}       A copy of the vnode
+   *
+  copyVnode: function (vnode) {
+    const vm = new Vue({
+      render: function () {
+        return vnode;
+      }
+    });
+    const copy = vm.$mount()._vnode;
+    vm.$destroy();
+    return copy;
+  },
+   */
+
+  /**
+   * Attempts a deep clone of the wrapper.vnode. Experimental,
+   * will hit a stack exceed max size error if vnode is too large.
+   * We don't want to mutate the original object, because it may be
+   * used again in the same test by another expect().
+   *
+   * @param  {object} wrapper  A Vue-Test-Utils wrapper
+   * @return {object}          A copy of the wrapper.vnode
+   */
+  cloneVnode: function (wrapper) {
+    if (wrapper && wrapper.vnode) {
+      let vnode;
+      try {
+        // vnode = this.copyVnode(wrapper.vnode);
+        vnode = _cloneDeep(wrapper.vnode);
+      } catch (err) {
+        console.log(err);
+      }
+      return vnode;
+    }
   }
 };
